Make Orbital Login Window Unclosable (scheme, window)
scheme:
added exit to list of parameters in calls to Window::new in SchemeMut.
Added conditions for handling window.resizable and window.exit events in
OrbitalScheme. If exit is false, modified to write 'e' to file in
SchemeMut

window:
Added exit bool to Window struct, as well as exit in parameters of
function new. Added handling to not render Exit if Exit flag is passed.
modified to write "e" to file if Exit flag is false