core: misc changes for GPU drivers
With the introduction of GPU drivers (`virtio-gpu`), the input that was
previously retrieved from `vesa:input` has been moved to its own scheme
`:input`.

This commit changes the file that orbital watches for events (from
`vesa:input` to `input:consumer`).

Also now the `display_path` is retrieved from the `DISPLAY` environment
variable instead. This variable is set in the init process.

Signed-off-by: Anhad Singh <andypythonappdeveloper@gmail.com>